Did some repackaging to make the classes look more organised
