--- src/main/java/duke/ui/TasklistUi.java
@@ -1,5 +1,0 @@
-package duke.ui;
-
-public class TasklistUi extends Ui {
-    
-}
