--- src/main/java/duke/ui/Ui.java
@@ -1,4 +1,0 @@
-package duke.ui;
-
-public abstract class Ui {
-}
