+++ src/main/java/tasklist/TaskList.java
@@ -1,14 +1,14 @@
-package duke;
+package tasklist;
 
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
 
-import duke.tasks.Deadline;
-import duke.tasks.Event;
-import duke.tasks.Task;
-import duke.tasks.ToDo;
+import tasks.Deadline;
+import tasks.Event;
+import tasks.Task;
+import tasks.ToDo;
 
 public class TaskList {
     private ArrayList<Task> tasks;
