+++ src/main/java/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package tasks;
 
 public class ToDo extends Task {
     public ToDo(String name) {
